Machine Learning/requirements.txt
flask
gunicorn
pandas
numpy
google-cloud-firestore
scikit-learn